requirements.txt
discord.py
python-dotenv
inquirer
discordLevelingSystem